Move SimplePie_Content_Type_Sniffer to namespace
