+++ src/Cache/Memcache.php
@@ -101,7 +101,7 @@ class Memcache implements Base
 				'prefix' => 'simplepie_',
 			),
 		);
-		$this->options = \SimplePie_Misc::array_merge_recursive($this->options, \SimplePie\Cache::parse_URL($location));
+		$this->options = \SimplePie\Misc::array_merge_recursive($this->options, \SimplePie\Cache::parse_URL($location));
 
 		$this->name = $this->options['extras']['prefix'] . md5("$name:$type");
 
@@ -117,7 +117,7 @@ class Memcache implements Base
 	 */
 	public function save($data)
 	{
-		if ($data instanceof \SimplePie)
+		if ($data instanceof \SimplePie\SimplePie)
 		{
 			$data = $data->data;
 		}
